requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.4.0
django-crontab==0.7.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
psycopg2-binary==2.9.9
PyJWT==2.8.0
sqlparse==0.5.0
typing_extensions==4.12.2
